refactor #192 : 인증 빌더 생성 및 페이지별로 적용하는 커스텀 훅 분리
