+++ src/features/login/login-page.tsx
@@ -1,34 +1,14 @@
-import { useEffect } from 'react';
-import { useNavigate } from 'react-router';
-
 import { css } from '@emotion/react';
 
 import Icon from '@/common/components/icon/icon';
-import { useAuthBuilder } from '@/common/hooks/use-auth-builder';
-import { useCheckQueryStrings } from '@/common/hooks/use-check-query-strings';
-import { useUserStore } from '@/store/user-auth';
+import { useLoginPageAuth } from '@/common/hooks/use-page-auth-service';
 
 import GoogleAuthButton from './components/custom-buttons/google-auth-button';
 
 import * as styles from './login-page.styles';
 
 function LoginPage() {
-  const { rollback: rollbackAuthBuilder, start: startAuthBuilder } = useAuthBuilder('Login');
-
-  const { isAuthenticated, userInfo } = useUserStore();
-  const navigate = useNavigate();
-  const isRollback = useCheckQueryStrings({ rollback: 'true' });
-
-  useEffect(() => {
-    // 인증 싸이클에 문제가 발생했기 때문에, 강제 로그아웃
-    if (isRollback) rollbackAuthBuilder();
-
-    // 최초 진입시, 로그인 진행
-    if (!isRollback && (!isAuthenticated || userInfo === null)) startAuthBuilder();
-
-    // 이미 로그인이 돼있을 경우
-    if (!isRollback && isAuthenticated && userInfo !== null) navigate('/upload');
-  }, [navigate, rollbackAuthBuilder, startAuthBuilder, isRollback, isAuthenticated, userInfo]);
+  useLoginPageAuth();
 
   return (
     <div css={styles.container}>
